Bugfix in sim reads

+++ src/merge_vcf/IntervallTree.h
@@ -22,7 +22,7 @@ private:
 	TNode * drl(TNode *&);
 	TNode * srr(TNode *&);
 	TNode * drr(TNode *&);
-	long overlap(breakpoint_str start, breakpoint_str stop,short type, std::pair<bool,bool> strands,SVS_Node * curr_svs);
+	int overlap(breakpoint_str start, breakpoint_str stop,short type, std::pair<bool,bool> strands,SVS_Node * curr_svs);
 	bool same_breakpoint(breakpoint_str first, breakpoint_str second,int max_dist);
 	void careful_screening(breakpoint_str &start, breakpoint_str& stop, short type, std::pair<bool, bool> strands, meta_data_str meta_info, TNode *&p);
 	long overlap_SNP(breakpoint_str start, SVS_Node * curr_svs);
